requirements.txt
flask==2.3.2
torch==2.0.1
torchvision==0.15.2
numpy==1.23.5
Pillow==9.5.0
pandas==1.5.3
open-clip-torch==2.19.1
